requirements.txt
django-admin-sortable2
pillow
django-nested-inlines
